Fix to update act_irrConsumption for not LimitAbstraction
For the option not LimitAbstraction, act_irrConsumption[2] and [3] are now updated. 
Previously, act_irrConsumption[2] and [3] were initialized at 0 and not updated.

Further details:
In the Soil module:

# availWaterInfiltration = water net from precipitation (- soil - interception - snow + snowmelt) + water for irrigation
availWaterInfiltration = availWaterInfiltration + self.var.act_irrConsumption[No] 

Since act_irrConsumption[No] was maintained at 0, the irrigation water was being withdrawn, but this water was not updating the available water for infiltration. This resulted in repeated near-maximum withdrawals while the soil water was only updated with water net from precipitation.