📤 Reservoir transfers
Basin transfers, inwards or outwards, can be performed using the reservoir transfers development, choosing waterbody ID 0 for the outside source/sink. Outwards transfers are added to industrial demand/use.

Giving reservoirs transfer water to associated Receiving reservoirs, daily as the associated fraction of live storage of the Giving reservoir, depending on available space in the Receiving reservoir. If the giving reservoir ID is 0 (Ocean), the fraction of live storage of the Receiving reservoir is used instead. Reservoirs can receive or give to several reservoirs.

reservoir transfers = True activates the Excel reservoir transfers.
Transfers are performed in the order they are in the Excel sheet.

Basin transfers are first applied with the daily rule through the netCDF if basin_transfers_daily_operations = True.