Interpolation Method - add method
Add peter's interpolation method to the downscaling2 function, so that the user can choose between 'spline' (currently used for CWatM public repo), 'bilinear' downscaling implemented by Sarah which does not depend on extent of model domain or 'peter' which is Peter's new interpolation function.
Probably we should rename 'peter' but I thought Peter might have a name idea for it.