♻️ Reservoir transfer 366-day
Reservoir transfers are now input with 366-day schedules. The previous version of a static daily transfer has been deprecated.

Reservoir transfers, inwards or outwards, are activated when reservoir_transfers = True in the text settings file. Choose the waterbody ID of the giving and receiving reservoirs and the amount, either as a fraction of live storage (values  ≤ 1)  or in cubic metres (values > 1). Waterbody ID 0 represents the infinite outside source/sink. Outward transfers are added to industrial demand/use.

Reservoir transfers occur daily, with Giving reservoirs transferring water to associated Receiving reservoirs. The amount of water transferred is determined by the associated fraction of live storage of the Giving reservoir, and the available space in the Receiving reservoir. If the Giving reservoir ID is 0 (Representing an outside source/sink), the water amount refers to the Receiving reservoir. Reservoirs can both receive and give to several reservoirs, and transfers are executed in the order they appear in the Excel sheet.